test(steami_screen): Add mock tests and rename screen.py to device.py.
