+++ lib/steami_screen/steami_screen/__init__.py
@@ -1,5 +1,5 @@
 from steami_screen.colors import rgb_to_gray4, rgb_to_rgb8, rgb_to_rgb565
-from steami_screen.screen import (
+from steami_screen.device import (
     BLACK,
     BLUE,
     DARK,
